test: fix basic functionality fork tests
